pyproject.toml
[project]
name = "projeto-catalogo-dados"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pymysql>=1.1.0",
    "pandas>=3.0.0",
    "pyyaml>=6.0.3",
]
